refactor(select): remove states and logic from interface.jsx
